+++ lib/tgchat/ext/td/td/telegram/UserId.h
@@ -1,5 +1,5 @@
 //
-// Copyright Aliaksei Levin (levlam@telegram.org), Arseny Smirnov (arseny30@gmail.com) 2014-2021
+// Copyright Aliaksei Levin (levlam@telegram.org), Arseny Smirnov (arseny30@gmail.com) 2014-2022
 //
 // Distributed under the Boost Software License, Version 1.0. (See accompanying
 // file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
@@ -20,7 +20,7 @@ class UserId {
   int64 id = 0;
 
  public:
-  static constexpr int64 MAX_USER_ID = (1ll << 40) - 1;
+  static constexpr int64 MAX_USER_ID = (static_cast<int64>(1) << 40) - 1;
 
   UserId() = default;
